refactor: Update logger and prisma imports for improved module structure
